extract sparql adapter to own file
